Move Sve2 GetColSums and GetAbsDxColSums to SimdSve2ColSums.cpp
Extract Sve2::GetColSums and Sve2::GetAbsDxColSums from
SimdSve2Statistic.cpp into a dedicated SimdSve2ColSums.cpp, and
register the new file in the VS2022 Sve2 project.
