+++ prj/vs2022/Sve2.vcxproj
@@ -43,6 +43,7 @@
     <ClCompile Include="..\..\src\Simd\SimdSve2BgrToLab.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2BgrToRgb.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2BgrToYuvV2.cpp" />
+    <ClCompile Include="..\..\src\Simd\SimdSve2ColSums.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2Conditional.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2Cpu.cpp" />
     <ClCompile Include="..\..\src\Simd\SimdSve2Detection.cpp" />
