+++ prj/vs2022/Sve2.vcxproj.filters
@@ -242,6 +242,9 @@
     <ClCompile Include="..\..\src\Simd\SimdSve2Statistic.cpp">
       <Filter>Sve2\Statistics</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Simd\SimdSve2ColSums.cpp">
+      <Filter>Sve2\Statistics</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Simd\SimdSve2StatisticMoments.cpp">
       <Filter>Sve2\Statistics</Filter>
     </ClCompile>
